refactor(cli): extract Login Flow Controller from login.ts
Extract the 604-line login.ts monolith into 3 focused modules,
following the design validated through the grilling skill:

1. login-shared.ts (~200 lines) — pure functions, types, constants.
   No I/O, no prompts, no process.exit(). Already tested (30 tests).

2. login-flow.ts (~280 lines) — 6 interactive flow functions returning
   FlowResult ({ action: "done" | "cancelled" | "error" }). Uses
   console.log directly to preserve the original incremental UX
   (headers appear before prompts, not after). Zero process.exit()
   calls. Prompts go through a DI singleton on prompt.ts.

3. login.ts (~100 lines) — thin handler wrappers that call flow
   functions and map FlowResult.action to exit codes. Re-exports all
   pure functions from login-shared.ts for backward compatibility.

4. prompt.ts — added setPromptDeps()/resetPromptDeps()/getPromptFn()/
   getPromptHiddenFn() DI singleton so login-flow.ts can be tested
   with mock prompts without affecting build-agent/plan-agent callers.

Key properties:
- No circular dependencies: login.ts → login-flow.ts → login-shared.ts
- All 16 process.exit() calls live in login.ts only
- Flow functions print directly (UX preserved — no buffered output)
- Cancelled action exits with 1 (matching original behavior)
- 1,288 tests pass, 0 fail. Typecheck + lint clean (239 files).
