+++ src/cli/handlers/login.ts
@@ -1,552 +1,63 @@
-import { readConfigFile, writeConfigFile } from "../../config/config-io.js";
-import { getConfigPath } from "../../config/loader.js";
-import type { Config } from "../../config/schema.js";
-import { detectProvider } from "../../providers/model-registry.js";
-import { prompt, promptHidden } from "../prompt.js";
-
-// Re-export for backward compatibility — tests import containsLiteralApiKey
-// from login.ts. The canonical home is now config-io.ts.
-export { containsLiteralApiKey } from "../../config/config-io.js";
-
-// ===== Types & Constants =====
-
-export interface LoginProviderInfo {
-	key: string;
-	name: string;
-	requiresApiKey: boolean;
-	envVar?: string;
-	defaultModel?: string;
-	getKeyUrl?: string;
-	note?: string;
-}
-
 /**
- * Providers available for login, ordered by popularity for the onboarding
- * picker. Keys match the `providers` object in config.yaml (see
- * `Config["providers"]` in schema.ts).
- */
-export const LOGIN_PROVIDERS: readonly LoginProviderInfo[] = [
-	{
-		key: "openai",
-		name: "OpenAI",
-		requiresApiKey: true,
-		envVar: "OPENAI_API_KEY",
-		defaultModel: "gpt-4o",
-		getKeyUrl: "https://platform.openai.com/api-keys",
-	},
-	{
-		key: "anthropic",
-		name: "Anthropic",
-		requiresApiKey: true,
-		envVar: "ANTHROPIC_API_KEY",
-		defaultModel: "claude-sonnet-4-20250514",
-		getKeyUrl: "https://console.anthropic.com/settings/keys",
-	},
-	{
-		key: "ollama",
-		name: "Ollama (Local)",
-		requiresApiKey: false,
-		defaultModel: "qwen3-coder",
-		note: "Runs locally — no API key needed. Install from ollama.com",
-	},
-	{
-		key: "ollamaCloud",
-		name: "Ollama (Cloud)",
-		requiresApiKey: true,
-		envVar: "OLLAMA_CLOUD_API_KEY",
-		defaultModel: "gpt-oss:120b-cloud",
-		getKeyUrl: "https://ollama.com",
-	},
-	{
-		key: "openrouter",
-		name: "OpenRouter",
-		requiresApiKey: true,
-		envVar: "OPENROUTER_API_KEY",
-		defaultModel: "openrouter/openai/gpt-4o",
-		getKeyUrl: "https://openrouter.ai/keys",
-	},
-	{
-		key: "opencode",
-		name: "OpenCode",
-		requiresApiKey: true,
-		envVar: "OPENCODE_API_KEY",
-		defaultModel: "opencode/big-pickle",
-		getKeyUrl: "https://opencode.ai",
-	},
-	{
-		key: "zai",
-		name: "Z.AI (Zhipu)",
-		requiresApiKey: true,
-		envVar: "ZAI_API_KEY",
-		defaultModel: "glm-4.7",
-		getKeyUrl: "https://open.bigmodel.cn/usercenter/apikeys",
-	},
-	{
-		key: "clinepass",
-		name: "ClinePass",
-		requiresApiKey: true,
-		envVar: "CLINE_API_KEY",
-		defaultModel: "cline-pass/glm-5.2",
-		getKeyUrl: "https://cline.bot",
-	},
-	{
-		key: "qwencloud",
-		name: "QwenCloud",
-		requiresApiKey: true,
-		envVar: "QWENCLOUD_API_KEY",
-		defaultModel: "qw/glm-5.2",
-		getKeyUrl: "https://home.qwencloud.com",
-	},
-];
-
-export interface ProviderStatus {
-	key: string;
-	name: string;
-	configured: boolean;
-	hasApiKey: boolean;
-	requiresApiKey: boolean;
-}
-
-// ===== Pure Functions (no I/O — directly testable) =====
-
-/** Find a login provider by key (case-insensitive). */
-export function findProvider(key: string): LoginProviderInfo | undefined {
-	const lower = key.toLowerCase();
-	return LOGIN_PROVIDERS.find((p) => p.key.toLowerCase() === lower);
-}
-
-/** Compute connection status for every provider from a config's providers object. */
-export function getProviderStatus(providers: Config["providers"] | undefined): ProviderStatus[] {
-	return LOGIN_PROVIDERS.map((p) => {
-		const providerConfig = providers?.[p.key as keyof NonNullable<typeof providers>];
-		const configured = !!providerConfig;
-		const hasApiKey = !!providerConfig?.apiKey;
-		return {
-			key: p.key,
-			name: p.name,
-			configured,
-			hasApiKey,
-			requiresApiKey: p.requiresApiKey,
-		};
-	});
-}
-
-export interface ApplyProviderOptions {
-	apiKey?: string;
-	baseUrl?: string;
-	defaultModel?: string;
-}
-
-/**
- * Merge a provider configuration into a raw config object (as read from the
- * config file). Preserves all existing keys (other providers, mcpServers,
- * skillDirectories, etc.) and only adds/overwrites the target provider +
- * optional defaultModel. Does NOT mutate the input.
- */
-export function applyProviderToConfig(
-	fileConfig: Record<string, unknown>,
-	providerKey: string,
-	options: ApplyProviderOptions
-): Record<string, unknown> {
-	const result: Record<string, unknown> = { ...fileConfig };
-
-	// Shallow-copy the providers object so we don't mutate the input
-	const existingProviders =
-		result.providers && typeof result.providers === "object"
-			? { ...(result.providers as Record<string, unknown>) }
-			: {};
-	result.providers = existingProviders;
-
-	const providers = existingProviders as Record<string, Record<string, unknown>>;
-	// Copy the existing provider entry to avoid mutation
-	const existing = { ...(providers[providerKey] ?? {}) };
-
-	const updated: Record<string, unknown> = { ...existing };
-	if (options.apiKey !== undefined) {
-		updated.apiKey = options.apiKey;
-	}
-	if (options.baseUrl !== undefined) {
-		updated.baseUrl = options.baseUrl;
-	}
-	providers[providerKey] = updated;
-
-	if (options.defaultModel !== undefined) {
-		result.defaultModel = options.defaultModel;
-	}
-
-	return result;
-}
-
-/** Format provider connection status as a human-readable string. */
-export function formatProviderStatus(providers: Config["providers"] | undefined): string {
-	const statuses = getProviderStatus(providers);
-	const lines: string[] = ["Provider Connection Status", "========================"];
-
-	for (const s of statuses) {
-		const icon = s.hasApiKey || (!s.requiresApiKey && s.configured) ? "●" : s.configured ? "◐" : "○";
-		let statusText: string;
-		if (s.requiresApiKey) {
-			statusText = s.hasApiKey ? "API key set" : s.configured ? "API key required" : "not configured";
-		} else {
-			statusText = s.configured ? "ready" : "ready (local)";
-		}
-		lines.push(`  ${icon} ${s.name.padEnd(16)} ${statusText}`);
-	}
-
-	return lines.join("\n");
-}
-
-// ===== Interactive Flows =====
-
-async function loginProvider(provider: LoginProviderInfo): Promise<void> {
-	const configPath = getConfigPath();
-	const fileConfig = await readConfigFile(configPath);
-
-	console.log(`\n🔐 ${provider.name} Login\n`);
-
-	if (provider.requiresApiKey) {
-		if (provider.getKeyUrl) {
-			console.log(`Get your API key at: ${provider.getKeyUrl}\n`);
-		}
-
-		let apiKey: string;
-		try {
-			apiKey = await promptHidden(`Enter your ${provider.name} API key: `);
-		} catch {
-			console.log("\n✗ Login cancelled.");
-			process.exit(1);
-		}
-		if (!apiKey) {
-			console.log("\n✗ No API key entered. Login cancelled.");
-			process.exit(1);
-		}
-
-		// Suggest a default model
-		let defaultModel: string | undefined;
-		if (provider.defaultModel) {
-			const existingDefault = fileConfig.defaultModel as string | undefined;
-			console.log(`\nSuggested default model: ${provider.defaultModel}`);
-			const confirm = await prompt("Set as default model? [Y/n]: ");
-			if (confirm.toLowerCase().startsWith("y") || confirm === "") {
-				defaultModel = provider.defaultModel;
-			} else if (!existingDefault) {
-				// No existing default model — must set one for a valid config
-				console.log(`(No default model set — using ${provider.defaultModel})`);
-				defaultModel = provider.defaultModel;
-			}
-		}
-
-		const updatedConfig = applyProviderToConfig(fileConfig, provider.key, {
-			apiKey,
-			defaultModel,
-		});
-		await writeConfigFile(configPath, updatedConfig);
-
-		console.log(`\n✓ Saved ${provider.name} API key to ${configPath}`);
-		if (defaultModel) {
-			console.log(`✓ Default model set to: ${defaultModel}`);
-		}
-	} else {
-		// No API key needed (e.g. Ollama local)
-		console.log(`${provider.note ?? "No API key needed."}\n`);
-
-		const existingBaseUrl =
-			(fileConfig.providers as Record<string, { baseUrl?: string }>)?.[provider.key]?.baseUrl ??
-			"http://localhost:11434";
-		const baseUrl = await prompt(`Base URL [${existingBaseUrl}]: `);
-		const finalBaseUrl = baseUrl || existingBaseUrl;
-
-		// Auto-set default model if none exists yet
-		let defaultModel: string | undefined;
-		if (provider.defaultModel) {
-			const existingDefault = fileConfig.defaultModel as string | undefined;
-			if (!existingDefault) {
-				defaultModel = provider.defaultModel;
-				console.log(`✓ Default model set to: ${defaultModel}`);
-			}
-		}
-
-		const updatedConfig = applyProviderToConfig(fileConfig, provider.key, {
-			baseUrl: finalBaseUrl,
-			defaultModel,
-		});
-		await writeConfigFile(configPath, updatedConfig);
-
-		console.log(`\n✓ Saved ${provider.name} configuration to ${configPath}`);
-	}
-
-	console.log(`\nLogin complete! Run \`tiny-agent chat\` to start.\n`);
-
-	if (provider.envVar) {
-		console.log(
-			`Tip: For better security, store the key in an environment variable instead:\n` +
-				`  1. Set apiKey: \${${provider.envVar}} in config.yaml\n` +
-				`  2. Export ${provider.envVar}=your-key in your shell profile.\n`
-		);
-	}
-
-	process.exit(0);
-}
-
-async function loginInteractive(): Promise<void> {
-	console.log("\n🔑 Tiny Agent — Provider Login\n");
-	console.log("Connect an LLM provider so you can start chatting.\n");
-
-	// Show current status
-	const configPath = getConfigPath();
-	const fileConfig = await readConfigFile(configPath);
-	const providers = fileConfig.providers as Config["providers"] | undefined;
-	console.log(formatProviderStatus(providers));
-	console.log();
-
-	// Show picker
-	console.log("Select a provider to configure:\n");
-	LOGIN_PROVIDERS.forEach((p, i) => {
-		const num = `${i + 1}.`.padStart(4);
-		const suffix = p.requiresApiKey ? `→ ${p.getKeyUrl ?? "API key required"}` : `→ ${p.note ?? "no API key needed"}`;
-		console.log(`  ${num} ${p.name.padEnd(16)} ${suffix}`);
-	});
-	console.log();
-
-	const choice = await prompt(`Enter choice (1-${LOGIN_PROVIDERS.length}): `);
-	const index = parseInt(choice, 10) - 1;
-
-	if (Number.isNaN(index) || index < 0 || index >= LOGIN_PROVIDERS.length) {
-		console.log("\n✗ Invalid choice. Login cancelled.");
-		process.exit(1);
-	}
-
-	const provider = LOGIN_PROVIDERS[index];
-	if (!provider) {
-		console.log("\n✗ Invalid choice. Login cancelled.");
-		process.exit(1);
-	}
-
-	await loginProvider(provider);
-}
-
-async function showLoginStatus(): Promise<void> {
-	const configPath = getConfigPath();
-	const fileConfig = await readConfigFile(configPath);
-	const providers = fileConfig.providers as Config["providers"] | undefined;
-
-	console.log(`\n${formatProviderStatus(providers)}\n`);
-	console.log("Run `tiny-agent login` to configure a provider.");
-	console.log("Run `tiny-agent login <provider>` to configure a specific provider directly.\n");
-	process.exit(0);
-}
-
-// ===== Logout: Pure Helper =====
-
-/**
- * Remove the `apiKey` from a provider's config entry, preserving `baseUrl`
- * and any other fields. Does NOT delete the provider entry itself — the
- * provider stays "configured" (e.g. with a custom baseUrl) but is
- * "disconnected" (no key). Does NOT mutate the input.
+ * login.ts — thin handler wrappers for the login/logout CLI commands.
  *
- * If the provider has no apiKey or no entry at all, the config is returned
- * unchanged (idempotent for the "already logged out" state).
+ * All flow logic lives in login-flow.ts (returns FlowResult with action
+ * only, zero process.exit() calls). All pure functions, types, and
+ * constants live in login-shared.ts. This file maps FlowResult.action
+ * to exit codes and re-exports the pure layer for backward compatibility.
+ *
+ * Architecture (ADR-016 decomposition):
+ *   login.ts → login-flow.ts → login-shared.ts
+ *   (handlers)   (flows)       (pure functions)
+ *   No circular dependencies — each layer depends only on the one below.
  */
-export function removeApiKeyFromConfig(
-	fileConfig: Record<string, unknown>,
-	providerKey: string
-): Record<string, unknown> {
-	const result: Record<string, unknown> = { ...fileConfig };
 
-	const existingProviders =
-		result.providers && typeof result.providers === "object"
-			? { ...(result.providers as Record<string, unknown>) }
-			: {};
-	result.providers = existingProviders;
+import {
+	type FlowResult,
+	loginInteractiveFlow,
+	loginProviderFlow,
+	logoutInteractiveFlow,
+	logoutProviderFlow,
+	showLoginStatusFlow,
+	showLogoutStatusFlow,
+} from "./login-flow.js";
+import { findProvider, LOGIN_PROVIDERS } from "./login-shared.js";
 
-	const providers = existingProviders as Record<string, Record<string, unknown>>;
-	const existing = providers[providerKey];
-	if (!existing || typeof existing !== "object") {
-		return result;
+// Re-export containsLiteralApiKey from config-io.ts (unchanged).
+export { containsLiteralApiKey } from "../../config/config-io.js";
+// Re-export FlowResult type for tests that may want to assert on it.
+export type { FlowResult } from "./login-flow.js";
+// Re-export for backward compatibility — tests and other modules import
+// pure functions, types, and constants from login.ts. The canonical home
+// is now login-shared.ts.
+export {
+	type ApplyProviderOptions,
+	applyProviderToConfig,
+	findProvider,
+	formatProviderStatus,
+	getProviderStatus,
+	isActiveProvider,
+	LOGIN_PROVIDERS,
+	type LoginProviderInfo,
+	type ProviderStatus,
+	removeApiKeyFromConfig,
+} from "./login-shared.js";
+
+// ===== Handler wrappers (exit-mappers only) =====
+
+function handleFlowResult(result: FlowResult): void {
+	if (result.action === "error" || result.action === "cancelled") {
+		process.exit(1);
 	}
-
-	// Shallow-copy and delete the apiKey key
-	const updated: Record<string, unknown> = { ...existing };
-	delete updated.apiKey;
-	providers[providerKey] = updated;
-
-	return result;
-}
-
-// ===== Logout: Interactive Flows =====
-
-/**
- * Check if the given provider key matches the provider that the current
- * `defaultModel` auto-detects to. Returns false if no defaultModel is set
- * or if detectProvider throws.
- */
-function isActiveProvider(providerKey: string, defaultModel: unknown): boolean {
-	if (typeof defaultModel !== "string" || !defaultModel) return false;
-	try {
-		return detectProvider(defaultModel) === providerKey;
-	} catch {
-		return false;
-	}
-}
-
-/**
- * Prompt the user to pick a new default model from the remaining configured
- * providers. If no other providers are configured, falls back to the Ollama
- * local default (qwen3-coder).
- */
-async function promptNewDefaultModel(
-	fileConfig: Record<string, unknown>,
-	loggedOutProviderKey: string
-): Promise<string | undefined> {
-	// Find providers that still have an apiKey (or don't require one)
-	const providers = (fileConfig.providers ?? {}) as Record<string, Record<string, unknown>>;
-	const candidates = LOGIN_PROVIDERS.filter((p) => {
-		if (p.key === loggedOutProviderKey) return false;
-		const cfg = providers[p.key];
-		if (!cfg) return false;
-		if (p.requiresApiKey) return !!cfg.apiKey;
-		return true; // no-key providers like Ollama local are always "ready"
-	});
-
-	if (candidates.length === 0) {
-		// No other providers configured — fall back to Ollama local default
-		const fallback = "qwen3-coder";
-		console.log(`\nNo other providers configured. Falling back to: ${fallback}`);
-		return fallback;
-	}
-
-	if (candidates.length === 1) {
-		const only = candidates[0];
-		if (only?.defaultModel) {
-			console.log(`\nSwitching default model to ${only.name}: ${only.defaultModel}`);
-			return only.defaultModel;
-		}
-	}
-
-	// Multiple candidates — show a picker
-	console.log("\nSelect a new default model:\n");
-	candidates.forEach((p, i) => {
-		const num = `${i + 1}.`.padStart(4);
-		const model = p.defaultModel ?? "(unknown)";
-		console.log(`  ${num} ${p.name.padEnd(16)} ${model}`);
-	});
-	console.log();
-
-	const choice = await prompt(`Enter choice (1-${candidates.length}): `);
-	const index = parseInt(choice, 10) - 1;
-
-	if (Number.isNaN(index) || index < 0 || index >= candidates.length) {
-		console.log("\n✗ Invalid choice. Keeping current default model.");
-		return undefined;
-	}
-
-	const selected = candidates[index];
-	return selected?.defaultModel;
-}
-
-async function logoutProvider(provider: LoginProviderInfo): Promise<void> {
-	// Refuse for providers that don't have an API key (e.g. Ollama local)
-	if (!provider.requiresApiKey) {
-		console.log(`\n${provider.name} has no API key to remove.`);
-		console.log(`Use 'tiny-agent login ${provider.key}' to reconfigure the base URL.\n`);
-		process.exit(0);
-	}
-
-	const configPath = getConfigPath();
-	const fileConfig = await readConfigFile(configPath);
-
-	// Check if the provider has an apiKey to remove
-	const providerConfig = (fileConfig.providers as Record<string, Record<string, unknown>>)?.[provider.key];
-	const hasApiKey = !!providerConfig?.apiKey;
-
-	if (!providerConfig) {
-		console.log(`\n${provider.name} is not configured. Nothing to log out.\n`);
-		process.exit(0);
-	}
-
-	if (!hasApiKey) {
-		console.log(`\n${provider.name} is configured but has no API key set. Already logged out.\n`);
-		process.exit(0);
-	}
-
-	// Remove the apiKey
-	let updatedConfig = removeApiKeyFromConfig(fileConfig, provider.key);
-
-	// Check if the logged-out provider was the active default model
-	const currentDefault = fileConfig.defaultModel;
-	if (isActiveProvider(provider.key, currentDefault)) {
-		console.log(`\n⚠️  The default model (${currentDefault}) uses ${provider.name}.`);
-		const newModel = await promptNewDefaultModel(updatedConfig, provider.key);
-		if (newModel) {
-			updatedConfig = { ...updatedConfig, defaultModel: newModel };
-			console.log(`✓ Default model set to: ${newModel}`);
-		}
-	}
-
-	await writeConfigFile(configPath, updatedConfig);
-
-	console.log(`\n✓ Removed ${provider.name} API key from ${configPath}`);
-	console.log(
-		`\nLogout complete! The ${provider.name} provider entry is preserved (baseUrl, etc.) but has no API key.\n`
-	);
 	process.exit(0);
 }
-
-async function logoutInteractive(): Promise<void> {
-	console.log("\n🔒 Tiny Agent — Provider Logout\n");
-
-	const configPath = getConfigPath();
-	const fileConfig = await readConfigFile(configPath);
-	const providers = fileConfig.providers as Config["providers"] | undefined;
-
-	// Show current status
-	console.log(formatProviderStatus(providers));
-	console.log();
-
-	// Filter to providers that have an API key set (can be logged out)
-	const providersWithKeys = LOGIN_PROVIDERS.filter((p) => {
-		if (!p.requiresApiKey) return false;
-		const cfg = providers?.[p.key as keyof NonNullable<typeof providers>];
-		return !!cfg?.apiKey;
-	});
-
-	if (providersWithKeys.length === 0) {
-		console.log("No providers have an API key set. Nothing to log out.\n");
-		process.exit(0);
-	}
-
-	// Show picker
-	console.log("Select a provider to log out:\n");
-	providersWithKeys.forEach((p, i) => {
-		const num = `${i + 1}.`.padStart(4);
-		console.log(`  ${num} ${p.name}`);
-	});
-	console.log();
-
-	const choice = await prompt(`Enter choice (1-${providersWithKeys.length}): `);
-	const index = parseInt(choice, 10) - 1;
-
-	if (Number.isNaN(index) || index < 0 || index >= providersWithKeys.length) {
-		console.log("\n✗ Invalid choice. Logout cancelled.");
-		process.exit(1);
-	}
-
-	const provider = providersWithKeys[index];
-	if (!provider) {
-		console.log("\n✗ Invalid choice. Logout cancelled.");
-		process.exit(1);
-	}
-
-	await logoutProvider(provider);
-}
-
-// ===== Main Handlers =====
 
 export async function handleLogin(args: string[]): Promise<void> {
 	const subCommand = args[0];
 
 	if (subCommand === "status") {
-		await showLoginStatus();
+		const result = await showLoginStatusFlow();
+		handleFlowResult(result);
 		return;
 	}
 
@@ -554,7 +65,8 @@ export async function handleLogin(args: string[]): Promise<void> {
 	if (subCommand) {
 		const provider = findProvider(subCommand);
 		if (provider) {
-			await loginProvider(provider);
+			const result = await loginProviderFlow(provider);
+			handleFlowResult(result);
 			return;
 		}
 		console.error(`Unknown provider: ${subCommand}`);
@@ -563,27 +75,16 @@ export async function handleLogin(args: string[]): Promise<void> {
 	}
 
 	// Interactive picker
-	await loginInteractive();
-}
-
-// ===== Logout: Status & Main Handler =====
-
-async function showLogoutStatus(): Promise<void> {
-	const configPath = getConfigPath();
-	const fileConfig = await readConfigFile(configPath);
-	const providers = fileConfig.providers as Config["providers"] | undefined;
-
-	console.log(`\n${formatProviderStatus(providers)}\n`);
-	console.log("Run `tiny-agent logout` to remove a provider's API key interactively.");
-	console.log("Run `tiny-agent logout <provider>` to log out a specific provider directly.\n");
-	process.exit(0);
+	const result = await loginInteractiveFlow();
+	handleFlowResult(result);
 }
 
 export async function handleLogout(args: string[]): Promise<void> {
 	const subCommand = args[0];
 
 	if (subCommand === "status") {
-		await showLogoutStatus();
+		const result = await showLogoutStatusFlow();
+		handleFlowResult(result);
 		return;
 	}
 
@@ -591,7 +92,8 @@ export async function handleLogout(args: string[]): Promise<void> {
 	if (subCommand) {
 		const provider = findProvider(subCommand);
 		if (provider) {
-			await logoutProvider(provider);
+			const result = await logoutProviderFlow(provider);
+			handleFlowResult(result);
 			return;
 		}
 		console.error(`Unknown provider: ${subCommand}`);
@@ -600,5 +102,6 @@ export async function handleLogout(args: string[]): Promise<void> {
 	}
 
 	// Interactive picker
-	await logoutInteractive();
+	const result = await logoutInteractiveFlow();
+	handleFlowResult(result);
 }
